CIO-65 Add src dir for babel
