+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "babel src -d lib",
-    "start": "node lib/index.js"
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
